requirements.txt
boto3==1.33.11
botocore==1.33.11
brotli==1.1.0
brotlicffi==1.1.0.0
chardet==5.2.0
cryptography==41.0.4
fastapi==0.106.0
pandas==2.1.4
pdf2image==1.16.3
protobuf==4.25.1
pyOpenSSL==23.3.0
PyPDF2==3.0.1
pytesseract==0.3.10
python_magic==0.4.27
sentence_transformers==2.2.2
simplejson==3.19.2
torch==2.0.1
urllib3_secure_extra==0.1.0
wordcloud==1.9.3
poppler=22.12.0
poppler-data=0.4.12